# Remove model constraints (and the rest) to make IK faster