--- a/main.py
+++ b/main.py
@@ -605,7 +605,7 @@ def main(sessionName, trialName, trial_id, cameras_to_use=['all'],
                 # Run IK tool. 
                 logging.info('Running Inverse Kinematics')
                 try:
-                    pathOutputIK = runIKTool(
+                    pathOutputIK, pathModelIK = runIKTool(
                         pathGenericSetupFile4IK, pathScaledModel, 
                         pathTRCFile4IK, outputIKDir)
                 except Exception as e:
@@ -623,7 +623,7 @@ def main(sessionName, trialName, trial_id, cameras_to_use=['all'],
         os.makedirs(outputJsonVisDir,exist_ok=True)
         outputJsonVisPath = os.path.join(outputJsonVisDir,
                                          trialName + '.json')
-        generateVisualizerJson(pathScaledModel, pathOutputIK,
+        generateVisualizerJson(pathModelIK, pathOutputIK,
                                outputJsonVisPath, 
                                vertical_offset=vertical_offset)