--- a/utilsOpenSim.py
+++ b/utilsOpenSim.py
@@ -163,11 +163,50 @@ def runIKTool(pathGenericSetupFile, pathScaledModel, pathTRCFile,
     pathOutputSetup =  os.path.join(
         pathOutputFolder, 'Setup_IK_' + IKFileName + '.xml')
     
-    # Setup IK tool.
+    # To make IK faster, we remove the patellas and their constraints from the
+    # model. Constraints make the IK problem more difficult, and the patellas
+    # are not used in the IK solution for this particular model. Since muscles
+    # are attached to the patellas, we also remove all muscles.
     opensim.Logger.setLevelString('error')
+    model = opensim.Model(pathScaledModel)
+    # Remove all actuators.                                         
+    forceSet = model.getForceSet()
+    i = 0
+    while i < forceSet.getSize():
+        forceSet.remove(i)
+    # Remove all constraints.
+    constraintSet = model.getConstraintSet()
+    i = 0
+    while i < constraintSet.getSize():
+        constraintSet.remove(i)        
+    # Remove patella bodies.
+    bodySet = model.getBodySet()
+    patella_bodies = ['patella_r', 'patella_l']
+    for patella in patella_bodies:
+        for i in range(bodySet.getSize()):        
+            c_body = bodySet.get(i).getName()
+            if c_body == patella:
+                bodySet.remove(i)
+                break
+    # Remove patellofemoral joints.
+    jointSet = model.getJointSet()
+    patellofemoral_joints = ['patellofemoral_r', 'patellofemoral_l']
+    for patellofemoral in patellofemoral_joints:
+        for i in range(jointSet.getSize()):
+            c_joint = jointSet.get(i).getName()
+            if c_joint == patellofemoral:
+                jointSet.remove(i)
+                break
+    # Print the model to a new file.
+    model.finalizeConnections
+    model.initSystem()
+    pathScaledModelWithoutPatella = pathScaledModel.replace('.osim', '_no_patella.osim')
+    model.printToXML(pathScaledModelWithoutPatella)   
+
+    # Setup IK tool.    
     IKTool = opensim.InverseKinematicsTool(pathGenericSetupFile)            
     IKTool.setName(IKFileName)
-    IKTool.set_model_file(pathScaledModel)          
+    IKTool.set_model_file(pathScaledModelWithoutPatella)          
     IKTool.set_marker_file(pathTRCFile)
     if timeRange:
         IKTool.set_time_range(0, timeRange[0])
@@ -180,7 +219,7 @@ def runIKTool(pathGenericSetupFile, pathScaledModel, pathTRCFile,
     command = 'opensim-cmd -o error' + ' run-tool ' + pathOutputSetup
     os.system(command)
     
-    return pathOutputMotion
+    return pathOutputMotion, pathScaledModelWithoutPatella
     
     
 # %% This function will look for a time window, of a minimum duration specified