@antoinefalisse 2024-07-03
@suhlrich and @carmichaelong, thanks for the feedback. Cleaner now and tested that I get the model out. Merging!